pyproject.toml
[project]
name = "agentflow"
version = "0.1.0"
description = "AI-native DevOps Automation Engine — 审批驱动，多云资源自动开通"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    # Lark (飞书)
    "lark-oapi>=1.4.0",
    # HTTP client
    "httpx>=0.28.0",
    # Settings
    "pydantic-settings>=2.6.0",
    # Database
    "sqlalchemy>=2.0.36",
    "aiosqlite>=0.20.0",
    # Alibaba Cloud SDKs
    "alibabacloud-devops20210625>=3.0.0",    # 云效 Flow
    "alibabacloud-alidns20150109>=3.0.0",    # DNS
    "alibabacloud-cas20200407>=3.0.0",       # SSL 证书
    "alibabacloud-rds20140815>=6.0.0",       # RDS
    "alibabacloud-r-kvstore20150101>=3.0.0", # Redis
    "alibabacloud-ecs20140526>=5.0.0",       # ECS
    "alibabacloud-slb20140515>=4.0.0",       # SLB
    "alibabacloud-sae20190506>=3.0.0",       # SAE
    "alibabacloud-bssopenapi20171214>=5.0.0",# BSS 询价
    "alibabacloud-tea-openapi>=0.3.0",
    "alibabacloud-credentials>=0.3.0",
    "oss2>=2.18.0",                          # OSS
    # Huawei Cloud
    "huaweicloudsdkcore>=3.1.0",
    "huaweicloudsdkecs>=3.1.0",
    # Tencent Cloud
    "tencentcloud-sdk-python-cvm>=3.0.0",
    # Scheduler
    "apscheduler>=3.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 100
